+++ RosBridgeClient/JObjectExtensions.cs
@@ -15,11 +15,9 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-
-using System;
 using Newtonsoft.Json.Linq;
 
-namespace RosBridgeClient
+namespace RosSharp.RosBridgeClient
 {
     public static class JObjectExtensions
     {
