+++ RosBridgeClient/Message.cs
@@ -15,10 +15,9 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-
 using Newtonsoft.Json;
 
-namespace RosBridgeClient
+namespace RosSharp.RosBridgeClient
 {
     public class Message
     {
