+++ RosBridgeClient/UrdfImporter.cs
@@ -15,7 +15,6 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-
 using System;
 using System.IO;
 using System.Collections.Generic;
@@ -23,7 +22,7 @@ using System.Threading;
 using System.Linq;
 using System.Xml.Linq;
 
-namespace RosBridgeClient
+namespace RosSharp.RosBridgeClient
 {
     public delegate void ReceiveEventHandler(ServiceReceiver sender, object ServiceResponse);
