+++ UrdfImporter/Robot.cs
@@ -19,7 +19,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Xml.Linq;
 
-namespace Urdf
+namespace RosSharp.Urdf
 {
     public class Robot
     {
